+++ js/data/game-start.js
@@ -3,8 +3,8 @@ import {render, enableBackButton, setScreen} from './../util';
 import screenHeader from './../screen/screen-header';
 import screenFooter from './../screen/screen-footer';
 import screenGame from './../screen/screen-game';
-import gameType from './game-type';
 import screenStat from './../screen/screen-stat';
+import getGameScreen from './get-game-screen';
 
 const gameStart = () => {
   const game = Object.assign({}, INITIAL_STATE);
@@ -25,13 +25,14 @@ const gameStart = () => {
     }
 
     const type = state.tasks[state.level].template;
+    const gameListener = getGameScreen(type);
 
     headerElement.innerHTML = screenHeader(state);
     gameElement.innerHTML = screenGame(state);
     footerElement.innerHTML = screenFooter(state);
 
     enableBackButton(wrapElement);
-    gameType[type].formListener(gameElement, state, changeTask);
+    gameListener.formListener(gameElement, state, changeTask);
   };
   setScreen(wrapElement);
   changeTask(game);
