+++ js/screen/screen-game.js
@@ -1,11 +1,12 @@
-import gameType from './../data/game-type';
+import getGameScreen from './../data/get-game-screen';
 
 export default (state) => {
   const task = state.tasks[state.level];
   const type = task.template;
+  const gameForm = getGameScreen(type);
 
   return `
   <section class="game">
     <p class="game__task">${task.text}</p>
-    ${gameType[type].formTemplate(task)}`;
+    ${gameForm.formTemplate(task)}`;
 };
